ITS WORKING!!!
Well, it was technically working before, but I was straight up assigning a char to an int for args like a idiot. That's how you get the ASCII value for things and have your program not work.

Booyah.
Also moved to O3 because paranoid and it runs wicked fast without O4 anyway.