A fixed optimization
This is a similiar optimization as the one that broke things, but I
figured out why. The isIn function checks to the bounds of t, which if
it's preallocated, is tournySize. However, there are no values there
yet, so it's just uninitialized memory, giving essentially random
values. If we allocate and initialize using -1, boom problem solved,
since it will always evaluate to false against the unset values. This
allows it to be a static array at compile time, and save a bunch of
runtime allocates for potentially few more comparison instructions.