OOP CUDA Conversion start
Began converting assignment 4 to use a class structure, decoupling the
highly-coupled components. Still a lot of work left to do (Ops and a lot
of the tree library), and a lot of errors to fix.
Once this is all done, the hope is to use CUDA programming to run on a
GPU, and evolve the parameters to populations.